solve problems

+++ 입국심사.cpp
@@ -1,21 +1,18 @@
-#include <iostream>
-#include <cstdio>
-#include <cstdlib>
-#include <cstring>
+// 문제가 개편 되었습니다. 이로 인해 함수 구성이 변경되어, 과거의 코드는 동작하지 않을 수 있습니다.
+// 새로운 함수 구성을 적용하려면 [코드 초기화] 버튼을 누르세요. 단, [코드 초기화] 버튼을 누르면 작성 중인 코드는 사라집니다.
+#include <string>
 #include <vector>
-#include <queue>
-#include <stack>
 
 using namespace std;
 
-int solution(int n, vector<int> times) {
+long long solution(int n, vector<int> times) {
     long long left = 1;
     long long right = 100000000000000;
-
+    
     while (left <= right) {
         long long sum = 0;
         long long mid = (left + right) / 2;
-        for (int i = 0; i < times.size(); i++) {
+        for (int i = 0; i<times.size(); i++) {
             sum += mid / times[i];
         }
         if (sum >= n) right = mid - 1;
